# docs/website/reference/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "additionalProperties": true,
        "properties": {
          "after_seq": {
            "minimum": 0,
            "type": "integer"
          },
          "agent_id": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "error": {
            "type": "string"
          },
          "event_seq": {
            "minimum": 0,
            "type": "integer"
          },
          "hint": {
            "type": "string"
          },
          "ok": {
            "const": false,
            "type": "boolean"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "JobResponse": {
        "additionalProperties": false,
        "properties": {
          "job": {
            "additionalProperties": true,
            "properties": {
              "created_at": {
                "format": "date-time",
                "type": "string"
              },
              "error": {
                "type": "string"
              },
              "id": {
                "type": "string"
              },
              "items": {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              "kind": {
                "type": "string"
              },
              "phase": {
                "type": "string"
              },
              "progress": {
                "additionalProperties": true,
                "type": "object"
              },
              "result": {
                "additionalProperties": true,
                "type": "object"
              },
              "status": {
                "enum": [
                  "queued",
                  "running",
                  "completed",
                  "failed"
                ],
                "type": "string"
              },
              "summary": {
                "type": "string"
              },
              "updated_at": {
                "format": "date-time",
                "type": "string"
              }
            },
            "required": [
              "id",
              "kind",
              "status",
              "phase",
              "progress",
              "summary",
              "items",
              "created_at",
              "updated_at"
            ],
            "type": "object"
          },
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "ok",
          "job"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "scheme": "bearer",
        "type": "http"
      },
      "CallbackToken": {
        "description": "Opaque capability secret path segment. Use placeholders in documentation and tests.",
        "in": "path",
        "name": "callback_token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Generated baseline for Holon's current HTTP control-plane surface. It is intentionally conservative: many response bodies remain JSON placeholders until the success/error envelope contracts stabilize.",
    "title": "Holon HTTP control-plane API",
    "version": "0.28.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/jobs/{job_id}": {
      "get": {
        "description": "Return a generic asynchronous job snapshot by id.",
        "operationId": "jobStatus",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            },
            "description": "Successful JSON response using a stable DTO schema."
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Client error JSON response."
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Server error JSON response."
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Job status",
        "tags": [
          "jobs"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Local runtime; port is configurable.",
      "url": "http://127.0.0.1:7878"
    }
  ],
...
}